Modified the evaluator to take crf as input

+++ main.py
@@ -33,7 +33,7 @@ def parse_arguments():
     parser.add_argument("--clip_value", dest="clip_value", type=float, default=5)
     parser.add_argument("--wdecay", dest="wdecay", type=float, default=0.00001)
     parser.add_argument("--step_size", dest="step_size", type=int, default=15)
-    parser.add_argument("--gamma", dest="gamma", type=int, default=10)
+    parser.add_argument("--gamma", dest="gamma", type=float, default=10)
     parser.add_argument("--validate_every", dest="validate_every", type=int, default=1)
 
     MODEL_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'models/')
